+++ changelog.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [0.2.3] [Github repo]
+### Fix
+- Serial device `is_reachable` method fix
+### Added
+- NEW BLE_DEVICE
 ## [0.2.2] - 2020-02-24
 ### Fix
 - WS_DEVICE cmd_nb fix for cmd method